+++ src/stdlib/stdlib_bisect.h
@@ -1,0 +1,8 @@
+#ifndef RADS_STDLIB_BISECT_H
+#define RADS_STDLIB_BISECT_H
+
+#include "../core/interpreter.h"
+
+void stdlib_bisect_register(void);
+
+#endif /* RADS_STDLIB_BISECT_H */